backend/requirements.txt
flask
flask_sqlalchemy
Flask-Login
Flask-Bcrypt
flask-migrate
flask-cors
flask-mail
Flask-JWT-Extended
pg8000
pandas
openpyxl
reportlab
flask-wtf
wtforms
wtforms-sqlalchemy
waitress
requests
python-dotenv
python-docx
python-dateutil
Werkzeug

